refactor sqs deveoper api into its own module
