+++ localstack-core/localstack/services/sqs/provider.py
@@ -1,5 +1,3 @@
-import copy
-import hashlib
 import json
 import logging
 import re
@@ -8,15 +6,12 @@ import time
 from collections.abc import Iterable
 from concurrent.futures.thread import ThreadPoolExecutor
 from itertools import islice
-from typing import Literal
 
 from botocore.utils import InvalidArnException
-from moto.sqs.models import BINARY_TYPE_FIELD_INDEX, STRING_TYPE_FIELD_INDEX
-from moto.sqs.models import Message as MotoMessage
 from werkzeug import Request as WerkzeugRequest
 
 from localstack import config
-from localstack.aws.api import CommonServiceException, RequestContext, ServiceException
+from localstack.aws.api import RequestContext, ServiceException
 from localstack.aws.api.sqs import (
     ActionNameList,
     AttributeNameList,
@@ -70,11 +65,8 @@ from localstack.aws.api.sqs import (
     Token,
     TooManyEntriesInBatchRequest,
 )
-from localstack.aws.protocol.parser import create_parser
-from localstack.aws.protocol.serializer import aws_response_serializer
 from localstack.aws.spec import load_service
 from localstack.config import SQS_DISABLE_MAX_NUMBER_OF_MESSAGE_LIMIT
-from localstack.http import Request, route
 from localstack.services.edge import ROUTER
 from localstack.services.plugins import ServiceLifecycleHook
 from localstack.services.sqs import constants as sqs_constants
@@ -84,6 +76,7 @@ from localstack.services.sqs.constants import (
     HEADER_LOCALSTACK_SQS_OVERRIDE_WAIT_TIME_SECONDS,
     MAX_RESULT_LIMIT,
 )
+from localstack.services.sqs.developer_api import SqsDeveloperApi
 from localstack.services.sqs.exceptions import (
     InvalidParameterValueException,
     MissingRequiredParameterException,
@@ -97,8 +90,10 @@ from localstack.services.sqs.models import (
     SqsStore,
     StandardQueue,
     sqs_stores,
+    to_sqs_api_message,
 )
 from localstack.services.sqs.utils import (
+    create_message_attribute_hash,
     decode_move_task_handle,
     generate_message_id,
     is_fifo_queue,
@@ -107,7 +102,6 @@ from localstack.services.sqs.utils import (
 )
 from localstack.services.stores import AccountRegionBundle
 from localstack.utils.aws.arns import parse_arn
-from localstack.utils.aws.request_context import extract_region_from_headers
 from localstack.utils.bootstrap import is_api_enabled
 from localstack.utils.cloudwatch.cloudwatch_util import (
     SqsMetricBatchData,
@@ -125,13 +119,6 @@ LOG = logging.getLogger(__name__)
 
 MAX_NUMBER_OF_MESSAGES = 10
 _STORE_LOCK = threading.RLock()
-
-
-class InvalidAddress(ServiceException):
-    code = "InvalidAddress"
-    message = "The address https://queue.amazonaws.com/ is not valid for this endpoint."
-    sender_fault = True
-    status_code = 404
 
 
 def assert_queue_name(queue_name: str, fifo: bool = False):
@@ -640,174 +627,6 @@ def check_fifo_id(fifo_id: str | None, parameter: str):
         )
 
 
-def get_sqs_protocol(request: Request) -> Literal["query", "json"]:
-    content_type = request.headers.get("Content-Type")
-    return "json" if content_type == "application/x-amz-json-1.0" else "query"
-
-
-def sqs_auto_protocol_aws_response_serializer(service_name: str, operation: str):
-    def _decorate(fn):
-        def _proxy(*args, **kwargs):
-            # extract request from function invocation (decorator can be used for methods as well as for functions).
-            if len(args) > 0 and isinstance(args[0], WerkzeugRequest):
-                # function
-                request = args[0]
-            elif len(args) > 1 and isinstance(args[1], WerkzeugRequest):
-                # method (arg[0] == self)
-                request = args[1]
-            elif "request" in kwargs:
-                request = kwargs["request"]
-            else:
-                raise ValueError(f"could not find Request in signature of function {fn}")
-
-            protocol = get_sqs_protocol(request)
-            return aws_response_serializer(service_name, operation, protocol)(fn)(*args, **kwargs)
-
-        return _proxy
-
-    return _decorate
-
-
-class SqsDeveloperEndpoints:
-    """
-    A set of SQS developer tool endpoints:
-
-    - ``/_aws/sqs/messages``: list SQS messages without side effects, compatible with ``ReceiveMessage``.
-    """
-
-    def __init__(self, stores=None):
-        self.stores = stores or sqs_stores
-
-    @route("/_aws/sqs/messages", methods=["GET", "POST"])
-    @sqs_auto_protocol_aws_response_serializer("sqs", "ReceiveMessage")
-    def list_messages(self, request: Request) -> ReceiveMessageResult:
-        """
-        This endpoint expects a ``QueueUrl`` request parameter (either as query arg or form parameter), similar to
-        the ``ReceiveMessage`` operation. It will parse the Queue URL generated by one of the SQS endpoint strategies.
-        """
-
-        if "x-amz-" in request.mimetype or "x-www-form-urlencoded" in request.mimetype:
-            # only parse the request using a parser if it comes from an AWS client
-            protocol = get_sqs_protocol(request)
-            operation, service_request = create_parser(
-                load_service("sqs", protocol=protocol)
-            ).parse(request)
-            if operation.name != "ReceiveMessage":
-                raise CommonServiceException(
-                    "InvalidRequest", "This endpoint only accepts ReceiveMessage calls"
-                )
-        else:
-            service_request = dict(request.values)
-
-        if not service_request.get("QueueUrl"):
-            raise QueueDoesNotExist()
-
-        try:
-            account_id, region, queue_name = parse_queue_url(service_request.get("QueueUrl"))
-        except ValueError:
-            LOG.error(
-                "Error while parsing Queue URL from request values: %s",
-                service_request.get,
-                exc_info=LOG.isEnabledFor(logging.DEBUG),
-            )
-            raise InvalidAddress()
-
-        if not region:
-            region = extract_region_from_headers(request.headers)
-
-        return self._get_and_serialize_messages(request, region, account_id, queue_name)
-
-    @route("/_aws/sqs/messages/<region>/<account_id>/<queue_name>")
-    @sqs_auto_protocol_aws_response_serializer("sqs", "ReceiveMessage")
-    def list_messages_for_queue_url(
-        self, request: Request, region: str, account_id: str, queue_name: str
-    ) -> ReceiveMessageResult:
-        """
-        This endpoint extracts the region, account_id, and queue_name directly from the URL rather than requiring the
-        QueueUrl as parameter.
-        """
-        return self._get_and_serialize_messages(request, region, account_id, queue_name)
-
-    def _get_and_serialize_messages(
-        self,
-        request: Request,
-        region: str,
-        account_id: str,
-        queue_name: str,
-    ) -> ReceiveMessageResult:
-        show_invisible = request.values.get("ShowInvisible", "").lower() in ["true", "1"]
-        show_delayed = request.values.get("ShowDelayed", "").lower() in ["true", "1"]
-
-        try:
-            store = SqsProvider.get_store(account_id, region)
-            queue = store.queues[queue_name]
-        except KeyError:
-            LOG.info(
-                "no queue named %s in region %s and account %s", queue_name, region, account_id
-            )
-            raise QueueDoesNotExist()
-
-        messages = self._collect_messages(
-            queue, show_invisible=show_invisible, show_delayed=show_delayed
-        )
-
-        return ReceiveMessageResult(Messages=messages)
-
-    def _collect_messages(
-        self, queue: SqsQueue, show_invisible: bool = False, show_delayed: bool = False
-    ) -> list[Message]:
-        """
-        Retrieves from a given SqsQueue all visible messages without causing any side effects (not setting any
-        receive timestamps, receive counts, or visibility state).
-
-        :param queue: the queue
-        :param show_invisible: show invisible messages as well
-        :param show_delayed: show delayed messages as well
-        :return: a list of messages
-        """
-        receipt_handle = "SQS/BACKDOOR/ACCESS"  # dummy receipt handle
-
-        sqs_messages: list[SqsMessage] = []
-
-        if show_invisible:
-            sqs_messages.extend(queue.inflight)
-
-        if isinstance(queue, StandardQueue):
-            sqs_messages.extend(queue.visible.queue)
-        elif isinstance(queue, FifoQueue):
-            if show_invisible:
-                for inflight_group in queue.inflight_groups:
-                    # messages that have been received are held in ``queue.inflight``, even for FIFO queues. however,
-                    # for fifo queues, messages that are in the same message group as messages that have been
-                    # received, are also considered invisible, and are held here in ``inflight_group.messages``.
-                    for sqs_message in inflight_group.messages:
-                        sqs_messages.append(sqs_message)
-
-            for message_group in queue.message_group_queue.queue:
-                # these are all messages of message groups that are visible
-                for sqs_message in message_group.messages:
-                    sqs_messages.append(sqs_message)
-        else:
-            raise ValueError(f"unknown queue type {type(queue)}")
-
-        if show_delayed:
-            sqs_messages.extend(queue.delayed)
-
-        messages = []
-
-        for sqs_message in sqs_messages:
-            message: Message = to_sqs_api_message(sqs_message, [QueueAttributeName.All], ["All"])
-            # these are all non-standard fields so we squelch the linter
-            if show_invisible:
-                message["Attributes"]["IsVisible"] = str(sqs_message.is_visible).lower()  # noqa
-            if show_delayed:
-                message["Attributes"]["IsDelayed"] = str(sqs_message.is_delayed).lower()  # noqa
-            messages.append(message)
-            message["ReceiptHandle"] = receipt_handle
-
-        return messages
-
-
 class SqsProvider(SqsApi, ServiceLifecycleHook):
     """
     LocalStack SQS Provider.
@@ -849,7 +668,6 @@ class SqsProvider(SqsApi, ServiceLifecycleHook):
         # default are limited to 500kb payload size by Werkzeug. we make sure we only *increase* the limit if it's
         # already set, and if it's already set to unlimited we leave it.
         from rolo import Request as RoloRequest
-        from werkzeug import Request as WerkzeugRequest
 
         # needed for the webserver integration (webservers create Werkzeug request objects)
         if WerkzeugRequest.max_form_memory_size is not None:
@@ -864,7 +682,7 @@ class SqsProvider(SqsApi, ServiceLifecycleHook):
 
     def on_before_start(self):
         query_api.register(ROUTER)
-        self._router_rules = ROUTER.add(SqsDeveloperEndpoints())
+        self._router_rules = ROUTER.add(SqsDeveloperApi())
         self._queue_update_worker.start()
         self._start_cloudwatch_metrics_reporting()
 
@@ -1159,7 +977,7 @@ class SqsProvider(SqsApi, ServiceLifecycleHook):
             MD5OfMessageBody=message["MD5OfBody"],
             MD5OfMessageAttributes=message.get("MD5OfMessageAttributes"),
             SequenceNumber=queue_item.sequence_number,
-            MD5OfMessageSystemAttributes=_create_message_attribute_hash(message_system_attributes),
+            MD5OfMessageSystemAttributes=create_message_attribute_hash(message_system_attributes),
         )
 
     def send_message_batch(
@@ -1205,7 +1023,7 @@ class SqsProvider(SqsApi, ServiceLifecycleHook):
                             MessageId=message.get("MessageId"),
                             MD5OfMessageBody=message.get("MD5OfBody"),
                             MD5OfMessageAttributes=message.get("MD5OfMessageAttributes"),
-                            MD5OfMessageSystemAttributes=_create_message_attribute_hash(
+                            MD5OfMessageSystemAttributes=create_message_attribute_hash(
                                 message.get("message_system_attributes")
                             ),
                             SequenceNumber=queue_item.sequence_number,
@@ -1259,7 +1077,7 @@ class SqsProvider(SqsApi, ServiceLifecycleHook):
             MD5OfBody=md5(message_body),
             Body=message_body,
             Attributes=self._create_message_attributes(context, message_system_attributes),
-            MD5OfMessageAttributes=_create_message_attribute_hash(message_attributes),
+            MD5OfMessageAttributes=create_message_attribute_hash(message_attributes),
             MessageAttributes=message_attributes,
         )
         if self._cloudwatch_dispatcher:
@@ -1819,34 +1637,6 @@ class SqsProvider(SqsApi, ServiceLifecycleHook):
             self._cloudwatch_dispatcher.shutdown()
 
 
-# Method from moto's attribute_md5 of moto/sqs/models.py, separated from the Message Object
-def _create_message_attribute_hash(message_attributes) -> str | None:
-    # To avoid the need to check for dict conformity everytime we invoke this function
-    if not isinstance(message_attributes, dict):
-        return
-    hash = hashlib.md5()
-
-    for attrName in sorted(message_attributes.keys()):
-        attr_value = message_attributes[attrName]
-        # Encode name
-        MotoMessage.update_binary_length_and_value(hash, MotoMessage.utf8(attrName))
-        # Encode data type
-        MotoMessage.update_binary_length_and_value(hash, MotoMessage.utf8(attr_value["DataType"]))
-        # Encode transport type and value
-        if attr_value.get("StringValue"):
-            hash.update(bytearray([STRING_TYPE_FIELD_INDEX]))
-            MotoMessage.update_binary_length_and_value(
-                hash, MotoMessage.utf8(attr_value.get("StringValue"))
-            )
-        elif attr_value.get("BinaryValue"):
-            hash.update(bytearray([BINARY_TYPE_FIELD_INDEX]))
-            decoded_binary_value = attr_value.get("BinaryValue")
-            MotoMessage.update_binary_length_and_value(hash, decoded_binary_value)
-        # string_list_value, binary_list_value type is not implemented, reserved for the future use.
-        # See https://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_MessageAttributeValue.html
-    return hash.hexdigest()
-
-
 def resolve_queue_location(
     context: RequestContext, queue_name: str | None = None, queue_url: str | None = None
 ) -> tuple[str, str | None, str]:
@@ -1869,106 +1659,6 @@ def resolve_queue_location(
             return context.account_id, context.region, queue_url
 
     return context.account_id, context.region, queue_name
-
-
-def to_sqs_api_message(
-    standard_message: SqsMessage,
-    attribute_names: AttributeNameList = None,
-    message_attribute_names: MessageAttributeNameList = None,
-) -> Message:
-    """
-    Utility function to convert an SQS message from LocalStack's internal representation to the AWS API
-    concept 'Message', which is the format returned by the ``ReceiveMessage`` operation.
-
-    :param standard_message: A LocalStack SQS message
-    :param attribute_names: the attribute name list to filter
-    :param message_attribute_names: the message attribute names to filter
-    :return: a copy of the original Message with updated message attributes and MD5 attribute hash sums
-    """
-    # prepare message for receiver
-    message = copy.deepcopy(standard_message.message)
-
-    # update system attributes of the message copy
-    message["Attributes"][MessageSystemAttributeName.ApproximateFirstReceiveTimestamp] = str(
-        int((standard_message.first_received or 0) * 1000)
-    )
-
-    # filter attributes for receiver
-    message_filter_attributes(message, attribute_names)
-    message_filter_message_attributes(message, message_attribute_names)
-    if message.get("MessageAttributes"):
-        message["MD5OfMessageAttributes"] = _create_message_attribute_hash(
-            message["MessageAttributes"]
-        )
-    else:
-        # delete the value that was computed when creating the message
-        message.pop("MD5OfMessageAttributes", None)
-    return message
-
-
-def message_filter_attributes(message: Message, names: AttributeNameList | None):
-    """
-    Utility function filter from the given message (in-place) the system attributes from the given list. It will
-    apply all rules according to:
-    https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/sqs.html#SQS.Client.receive_message.
-
-    :param message: The message to filter (it will be modified)
-    :param names: the attributes names/filters
-    """
-    if "Attributes" not in message:
-        return
-
-    if not names:
-        del message["Attributes"]
-        return
-
-    if QueueAttributeName.All in names:
-        return
-
-    for k in list(message["Attributes"].keys()):
-        if k not in names:
-            del message["Attributes"][k]
-
-
-def message_filter_message_attributes(message: Message, names: MessageAttributeNameList | None):
-    """
-    Utility function filter from the given message (in-place) the message attributes from the given list. It will
-    apply all rules according to:
-    https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/sqs.html#SQS.Client.receive_message.
-
-    :param message: The message to filter (it will be modified)
-    :param names: the attributes names/filters (can be 'All', '.*', '*' or prefix filters like 'Foo.*')
-    """
-    if not message.get("MessageAttributes"):
-        return
-
-    if not names:
-        del message["MessageAttributes"]
-        return
-
-    if "All" in names or ".*" in names or "*" in names:
-        return
-
-    attributes = message["MessageAttributes"]
-    matched = []
-
-    keys = [name for name in names if ".*" not in name]
-    prefixes = [name.split(".*")[0] for name in names if ".*" in name]
-
-    # match prefix filters
-    for k in attributes:
-        if k in keys:
-            matched.append(k)
-            continue
-
-        for prefix in prefixes:
-            if k.startswith(prefix):
-                matched.append(k)
-            break
-    if matched:
-        message["MessageAttributes"] = {k: attributes[k] for k in matched}
-    else:
-        message.pop("MessageAttributes")
 
 
 def extract_message_count_from_headers(context: RequestContext) -> int | None:
