+++ tests/unit/services/sqs/test_sqs.py
@@ -4,8 +4,8 @@ import localstack.services.sqs.exceptions
 import localstack.services.sqs.models
 from localstack.services.sqs import provider
 from localstack.services.sqs.constants import DEFAULT_MAXIMUM_MESSAGE_SIZE
-from localstack.services.sqs.provider import _create_message_attribute_hash
 from localstack.services.sqs.utils import (
+    create_message_attribute_hash,
     guess_endpoint_strategy_and_host,
     is_sqs_queue_url,
     parse_queue_url,
@@ -20,7 +20,7 @@ def test_sqs_message_attrs_md5():
             "DataType": "Number",
         }
     }
-    md5 = _create_message_attribute_hash(msg_attrs)
+    md5 = create_message_attribute_hash(msg_attrs)
     assert md5 == "235c5c510d26fb653d073faed50ae77c"
